refactor: created seperate PayoutItem component
